+++ packages/identity/src/index.ts
@@ -1,4 +1,4 @@
-import { createRootCA, type RootCA } from './generateRootCA'
+import { createRootCA, type RootCA } from './createRootCA'
 import {
   extractPubKey,
   parseCertificate,
@@ -9,10 +9,10 @@ import {
   pubKeyFromCsr,
 } from './extractPubKey'
 import { verifyUserCert } from './verifyUserCertificate'
-import { verifySignature } from './verification'
+import { verifySignature } from './verifySignature'
 import { sign } from './sign'
-import { createUserCsr, type UserCsr } from './requestCertificate'
-import { createUserCert, type UserCert } from './generateUserCertificate'
+import { createUserCsr, type UserCsr } from './createUserCsr'
+import { createUserCert, type UserCert } from './createUserCert'
 import {
   loadPrivateKey,
   loadCSR,
@@ -24,6 +24,8 @@ import {
   hexStringToArrayBuffer,
   arrayBufferToHexString,
   getReqFieldValue,
+  certificateByUsername,
+  pubKeyMatch,
 } from './common'
 import configCrypto from './config'
 import {
@@ -62,6 +64,8 @@ export {
   CertFieldsTypes,
   hexStringToArrayBuffer,
   arrayBufferToHexString,
+  certificateByUsername,
+  pubKeyMatch,
 }
 export { configCrypto }
 export { setupCrypto }
