+++ packages/identity/src/test/helpers.ts
@@ -1,8 +1,8 @@
 import { Time, setEngine, CryptoEngine } from 'pkijs'
 import { Crypto } from '@peculiar/webcrypto'
-import { createRootCA, type RootCA } from '../generateRootCA'
-import { createUserCert, type UserCert } from '../generateUserCertificate'
-import { createUserCsr, type UserCsr } from '../requestCertificate'
+import { createRootCA, type RootCA } from '../createRootCA'
+import { createUserCert, type UserCert } from '../createUserCert'
+import { createUserCsr, type UserCsr } from '../createUserCsr'
 import config from '../config'
 
 export const userData = {
