+++ packages/identity/src/verifySignature.ts
